Move keys to 1 file
