+++ cmd/lib/commandService.js
@@ -1,3 +1,5 @@
+// TODO: remove this file when done refactoring
+
 const fs = require('fs')
 const VError = require('verror')
 const {
@@ -16,7 +18,7 @@ const dockerComposeCmd = `docker-compose -f ${getDockerComposeStartFile()} -f ${
 const nodeCmd = 'exec cudos-node cudos-noded '
 const nodeMultiCmd = 'exec cudos-node sh -c '
 
-const doDocker = function (cmd) {
+const doDocker = function(cmd) {
   const childResult = spawnSync(cmd, {
     stdio: 'inherit',
     shell: true
@@ -26,24 +28,24 @@ const doDocker = function (cmd) {
   }
 }
 
-const execute = function (arg) {
+const execute = function(arg) {
   const cmd = dockerComposeCmd + arg
   doDocker(cmd)
 }
 
-const executeNode = function (arg) {
+const executeNode = function(arg) {
   execute(nodeCmd + arg)
 }
 
-const MultiExecuteNode = function (arg) {
+const MultiExecuteNode = function(arg) {
   execute(nodeMultiCmd + `'${arg}'`)
 }
 
-const stopNode = function () {
+const stopNode = function() {
   execute(' down')
 }
 
-const startNode = function (inBackground) {
+const startNode = function(inBackground) {
   if (inBackground) {
     execute(' up --build -d')
   } else {
@@ -51,15 +53,15 @@ const startNode = function (inBackground) {
   }
 }
 
-const keysListNode = function () {
+const keysListNode = function() {
   executeNode('keys list')
 }
 
-const addAccountNode = function (name) {
+const addAccountNode = function(name) {
   MultiExecuteNode(`cudos-noded keys add ${name} && ` + transferTokensByNameCommand('faucet', name, '1000000000000000000'))
 }
 
-const deleteAccountNode = function (name, confirm) {
+const deleteAccountNode = function(name, confirm) {
   if (confirm) {
     executeNode(`keys delete ${name} --yes`)
   } else {
@@ -67,11 +69,11 @@ const deleteAccountNode = function (name, confirm) {
   }
 }
 
-const fundAccountNode = function (name, amount) {
+const fundAccountNode = function(name, amount) {
   MultiExecuteNode(transferTokensByNameCommand('faucet', name, amount))
 }
 
-const compile = function () {
+const compile = function() {
   const projectRootPath = getProjectRootPath()
   const compileCmd = `docker run --rm -v "${projectRootPath}":/code  --mount type=volume,source="contracts_cache",target=/code/target  --mount type=volume,source=registry_cache,target=/usr/local/cargo/registry cosmwasm/workspace-optimizer:${optimizerVer}`
 
@@ -82,7 +84,7 @@ const compile = function () {
   doDocker(compileCmd)
 }
 
-function transferTokensByNameCommand (fromName, toName, amount) {
+function transferTokensByNameCommand(fromName, toName, amount) {
   return `cudos-noded tx bank send ${fromName} $(cudos-noded keys show ${toName} -a) ${amount}acudos ` +
   '--chain-id cudos-network --yes'
 }
